Get chat parser working
It uses a regex to parse the usernames and batches of text they
typed. Usernames are passed to a unique userArray (which also gets
mapped to a formArray for character selection), while text is simply
added to the frames straight up.

Still need to figure out how to do the autocompletes & fix emoji
support/handling.