Implement chara select & fix encoding bug
This mimics the one in the application (named groups, searchable).
Might be a good idea to have the selected chara's name appear beside
so that the ID name doesn't become a weird thing.

Also it seems like the poseIDs are not retrievable from the site for
some fucked reason so I'll have to manually retrieve some. Adding
them into the character_ids file for now.

Emojis are now encodable which allows for some... interesting name
results... turns out the application also supports emojis in-text.